Add MergeSorter classes to sort.js
This commit adds two new sorter classes to the sort.js module.

The first class, MergeSorter, presents all components of the classic
Merge Sort algorithm. It shows the splitting of the list into smaller
sublists, the merging of said sublists, and the recursive nature of
the algorithm. It is meant for curious users only, for it is visually
dangerous (i.e may require a seizure warning).

The second class, SimplifiedMergeSorter, presents a simpler version of
the algorithm that hides all the merging and extra space usage and
makes it SEEM as if it were in-place (though it is not, but it would
be a good exercise for later).

Moreover, some extra functionality has been added to the Snapshot
class for ease of manipulation.