Update createScrambledRangeList function algorithm
This commit changes the current implementation of the algorithm that
creates an array containing a scrambled range of values. By scrambled
range, we specify that, for a range of integers from 1 to a, where a
is an arbitrary positive integer:

    - Every integer between 1 to a is included in the scrambled list.
    - No element of the scrambled list occurs more than once.

Previous algorithms were quit slow on big values, slowing the
execution of the sorting presentation down. This created bad user
experiences on our test cases. One of them had an best-case runtime
of Theta(n^2), which was quite slow on big list sizes. Another
required less operations, but these operations were still quite
expensive.

The new algorithm builds a list from an ever decreasing range of
values rather than creating a list and doing swaps. After many tests,
the algorithm works much faster, creating a list of size 100
almost instantly compared to previous algorithms. The correctness of
this algorithm remains to be proven, but all test cases have been
satisfied without error.