Add QuickSorter class to sort.js
This commit introduces the Quicksort algorithm sorter class. It
recursively sorts the list "in-place," using the classic Hoare
partition scheme created by the inventor of the algorithm.

Many other partition schemes and pivot strategies can be used and
added later, but currently, this is the most basic form of the
algorithm.