Add clear functionality for canvas partitions
This commit adds new clear functions/methods for canvas partitions.
Previously, it was only possible to clear the whole canvas. But since
we are able to partition the canvas, we should be able to
clear sections of the canvas right from the Partition class interface.

Two functions are provided. One is a method added to the Partition
class definition. It clears the canvas section enclosed by the
partition. The second is a function that works on a List of
partitions. It clears each partition in it.